+++ src/index.ts
@@ -5,12 +5,8 @@ export { Kernel as default } from './client';
 export { type Uploadable, toFile } from './core/uploads';
 export { APIPromise } from './core/api-promise';
 export { Kernel, type ClientOptions } from './client';
-export {
-  BrowserRouteCache,
-  type BrowserFetchInit,
-  type BrowserRoute,
-  type BrowserRoutingOptions,
-} from './lib/browser-routing';
+export { type BrowserFetchInit } from './lib/browser-fetch';
+export { BrowserRouteCache, type BrowserRoute, type BrowserRoutingOptions } from './lib/browser-routing';
 export { PagePromise } from './core/pagination';
 export {
   KernelError,
